+++ tdc/chemutils/__init__.py
@@ -3,3 +3,4 @@ from .evaluator import *
 from .molconvert import * 
 from .oracle import * 
 from ._smiles2pubchem import smiles2pubchem
+from .molfilter import *
